#182 - Move deprecated code
