+++ src/deprecated/nimble_darma_utils.cc
