+++ src/deprecated/nimble_darma_utils.h
